% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metid.R
\name{addMetid}
\alias{addMetid}
\title{Compare against a MS2 library}
\usage{
addMetid(x, metids)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file.}

\item{metids}{\code{data.frame}}
}
\value{
\code{list} Same as x, but additionally contains columns from metids
}
\description{
\code{addMetid} allows to matches Cluster from different ion modes against each
other. Cluster are matched by retention time and then checked if they share
a common neutral mass.
}
