% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metid.R
\name{annotateMz}
\alias{annotateMz}
\title{function to perform MS1 annotation}
\usage{
annotateMz(
  x,
  ms1Library,
  tolerance = 0,
  ppm = 0,
  rtimeTolerance = 10,
  matchAdduct = FALSE
)
}
\description{
function to perform MS1 annotation
}
