% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metid.R
\name{compareSpectraExternal}
\alias{compareSpectraExternal}
\title{Compare against a MS2 library}
\usage{
compareSpectraExternal(
  x,
  librarySpectra,
  treshold = 0.7,
  tolerance = 0,
  ppm = 0,
  ...
)
}
\arguments{
\item{x}{\code{Spectra} Spectra for which library matching shall be performed}

\item{librarySpectra}{\code{Spectra} Spectra object containing library}

\item{treshold}{\code{numeric} minimum spectral similarity, default is 0.7 for ndotproduct}

\item{tolerance}{\code{numeric} absolute tolerance for matching of peaks}

\item{ppm}{\code{numeric} relative tolerance for matching of peaks}
}
\value{
\code{data.frame} with the results
}
\description{
\code{compareSpectraExternal} compares measured spectra stored in a <code>Spectra
</code> object against a library also stored in a <code>Spectra</code>
object. Difference to \code{compareSpectraInHouse} is that RT is ignored.
}
\examples{

}
