% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metid.R
\name{compareSpectraInHouse}
\alias{compareSpectraInHouse}
\title{Compare against a in-house MS2 library}
\usage{
compareSpectraInHouse(
  x,
  librarySpectra,
  treshold = 0.7,
  tolerance = 0,
  ppm = 0,
  rtOffset = 0,
  rtimeTolerance = 10,
  ...
)
}
\arguments{
\item{x}{\code{Spectra} Spectra for which library matching shall be performed}

\item{librarySpectra}{\code{Spectra} Spectra object containing library}

\item{treshold}{\code{numeric} minimum spectral similarity, default is 0.7 for ndotproduct}

\item{tolerance}{\code{numeric} absolute tolerance for matching of peaks}

\item{ppm}{\code{numeric} relative tolerance for matching of peaks}

\item{rtOffset}{\code{numeric} known offset between measurement and library RT}

\item{rtimeTolerance}{\code{numeric} tolerance for retention time search}
}
\value{
\code{data.frame} with the results
}
\description{
\code{compareSpectraInHouse} compares measured spectra stored in a <code>Spectra
</code> object against a library also stored in a <code>Spectra</code>
object. Difference to \code{compareSpectraExternal} is that RT is used for
additional filtering.
}
\examples{

}
