% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metid.R
\name{compareSpectraLibrary}
\alias{compareSpectraLibrary}
\title{Compare against a MS2 library}
\usage{
compareSpectraLibrary(
  x,
  ms2library,
  tolerance = 0,
  ppm = 0,
  rtOffset = 0,
  rtimeTolerance = Inf,
  plot = FALSE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{x}{\code{Spectra} Spectra for which library matching shall be performed}

\item{ms2library}{\code{Spectra} Spectra object containing library}

\item{tolerance}{\code{numeric} absolute tolerance for matching of peaks}

\item{ppm}{\code{numeric} relative tolerance for matching of peaks}

\item{rtOffset}{\code{numeric} known offset between measurement and library RT}

\item{rtimeTolerance}{\code{numeric} tolerance for retention time search}
}
\value{
\code{data.frame} with the results
}
\description{
\code{compareSpectraLibrary} compares measured spectra stored in a <code>Spectra
</code> object against a library also stored in a <code>Spectra</code>
object. By default rtimeTolerance is set to Inf, if RT matching shall be
performed a different value in seconds needs to be set.
}
\examples{

}
