% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms2.R
\name{correctRtime}
\alias{correctRtime}
\title{Correction of retention time in Spectra objects}
\usage{
correctRtime(x, spectra)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file containing grouped
MS1 cluster}

\item{spectra}{\code{Spectra} Spectra object containing MS1 or MS2 spectra for
which the RT shall be corrected}
}
\description{
\code{correctRtime} allows to correct the retention time in <code>Spectra</code>
objects. It uses the RT from the MS1 data supplied from a .gda read data.
<code>Spectra</code> can be of either MS1 or MS2 level
}
