% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metabolights.R
\name{create_assay_file}
\alias{create_assay_file}
\title{Reconstructing Isotope Pattern from MS1 data}
\usage{
create_assay_file(x, study_id)
}
\arguments{
\item{peaks}{\code{list} List with data read from .gda file containing individual
MS1 peaks}

\item{cluster}{\code{list}List with data read from .gda file containing grouped
MS1 cluster}
}
\value{
\code{Spectra} Returns a Spectra object with the reconstructed MS1 isotope
pattern
}
\description{
\code{reconstructIsoPattern} Genedata does not allow to export MS1 isotope pattern
directly. They have to be reconstructed from the Peak and Cluster data
}
\examples{


}
\author{
Michael Witting
}
