% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gnps.R
\name{create_gnps_files}
\alias{create_gnps_files}
\title{Create input for GNPS feature based molecular networking

This functions takes data from MS1 cluster as well as MS2 spectra that 
contain CLUSTER_ID in the metadata. Two output files are generated. The first
 one a .mgf file with all the metadata required for linking the MS1 and MS2 
 data. The second file contains the MS1 table in a format similar to XCMS3.
 Both the .mgf file and the feature table can be used with the XCMS3 branch 
 of the GNPS feature based molecular networking}
\usage{
create_gnps_files(ms_data, row_anno, ms2_spectra_id, path_to_save)
}
\arguments{
\item{ms_data}{Data frame with the intensity or peak area values}

\item{row_anno}{Data frame with the row annotations}

\item{ms2_spectra_id}{Spectra object with the MS2 spectra, CLUSTER_ID column in metadata is required}

\item{path_to_save}{Path where the results files shall be saved}
}
\description{
Create input for GNPS feature based molecular networking

This functions takes data from MS1 cluster as well as MS2 spectra that 
contain CLUSTER_ID in the metadata. Two output files are generated. The first
 one a .mgf file with all the metadata required for linking the MS1 and MS2 
 data. The second file contains the MS1 table in a format similar to XCMS3.
 Both the .mgf file and the feature table can be used with the XCMS3 branch 
 of the GNPS feature based molecular networking
}
