% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gnps.R
\name{createGnpsFiles}
\alias{createGnpsFiles}
\title{Reformat .gda and MS2 spectra for GNPS FBMN}
\usage{
createGnpsFiles(x, spectra)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file containing grouped
MS1 cluster}

\item{spectra}{\code{Spectra}MSnbase Spectra object containing MS2 spectra to be
matched with the MS1 data}
}
\value{
A \code{list} with the reformated feature table in XCMS3 format as first
element and a reformated \code{Spectra} as second element.
}
\description{
\code{createGnpsFiles} reformats the Gendata data matrix and spectra to be consistent
with the requirement of GNPS FBMN. The resulting data can be fed into the
XCMS3 workflow.
}
\examples{


}
\author{
Michael Witting
}
