% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metabolights.R
\name{createMtblsMetaboliteFile}
\alias{createMtblsMetaboliteFile}
\title{Writing template for Metabolights metabolite file}
\usage{
createMtblsMetaboliteFile(x, study_id)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file}

\item{study_id}{\code{character} Study ID from Metabolights or place holder
(default is MTBLSXXXX)}
}
\description{
\code{createMtblsMetaboliteFile} Exports data read from a .gda file to a Metabolights
MAF file. A study ID is required to generate consistent files. To be
used this function requires the following column annotations to be present:
"Source \link{C}", "Sample \link{C}", "Assay \link{C}", "File \link{C}", "Sample Type \link{C}"
}
\details{
#' @title Writing template for Metabolights sample file
#'
#' @description
#'
#' \code{createMtblsSampleFile} Exports data read from a .gda file to a Metabolights
#'    sample file. A study ID is required to generate consistent files. To be
#'    used this function requires the following column annotations to be present:
#'    "Source \link{C}", "Sample \link{C}", "Assay \link{C}", "File \link{C}", "Sample Type \link{C}"
#'
#' @param x \code{list} List with data read from .gda file
#' @param study_id \code{character} Study ID from Metabolights or place holder
#'     (default is MTBLSXXXX)
#' @param study_factors \code{character} Potentially present study factors
#' @param metadata \code{list} Additional metadata, such a organism etc.
#'#'
#' @author Michael Witting
#'
#' @export
#'
#' @examples
#'
#'
create_assay_file <- function(x, study_id) {

}
}
\examples{


}
\author{
Michael Witting
}
