% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metabolights.R
\name{createMtblsSampleFile}
\alias{createMtblsSampleFile}
\title{Writing template for Metabolights sample file}
\usage{
createMtblsSampleFile(
  x,
  study_id = "MTBLSXXXX",
  study_factors = NA,
  metadata = list(organism_characteristics = NA, organism_source_ref = NA,
    organism_term_access = NA, organism_part_characteristics = NA,
    organism_part_source_ref = NA, organism_part_term_access = NA,
    variant_characteristics = NA, variant_source_ref = NA, variant_term_access = NA,
    protocol_ref = NA)
)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file}

\item{study_id}{\code{character} Study ID from Metabolights or place holder
(default is MTBLSXXXX)}

\item{study_factors}{\code{character} Potentially present study factors}

\item{metadata}{\code{list} Additional metadata, such a organism etc.
#'}
}
\description{
\code{createMtblsSampleFile} Exports data read from a .gda file to a Metabolights
sample file. A study ID is required to generate consistent files. To be
used this function requires the following column annotations to be present:
"Source \link{C}", "Sample \link{C}", "Assay \link{C}", "File \link{C}", "Sample Type \link{C}"
}
\examples{


}
\author{
Michael Witting
}
