% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metid.R
\name{.isolate_metadata}
\alias{.isolate_metadata}
\title{helper function to isolate metadata from different backends}
\usage{
.isolate_metadata(x)
}
\description{
helper function to isolate metadata from different backends
}
