% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ms2.R
\name{exportMgf}
\alias{exportMgf}
\title{Exporting Spectra as .mgf file}
\usage{
exportMgf(splist, con, COM = NULL, TITLE = NULL)
}
\arguments{
\item{splist}{\code{Spectra} Spectra object containing spectra to be exported}

\item{con}{\code{character} File name for export to .mgf}

\item{COM}{\code{character} Optional comment}

\item{TITLE}{\code{character} Optional title}
}
\description{
\code{exportMgf} This function allows to export <code>Spectra</code> object to .mgf
files. If the <code>Spectra</code> object contains the CLUSTER_ID and RAWFILE variables
this is additionally exported as additional metadata fields.
}
\examples{


}
\author{
Michael Witting
}
