% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms1.R
\name{fillMs1Spectra}
\alias{fillMs1Spectra}
\title{Fill with empty MS2 spectra}
\usage{
fillMs1Spectra(x, spectra)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file containing grouped
MS1 cluster}

\item{spectra}{\code{Spectra} Spectra object containing MS1 spectra for
which the RT shall be corrected}
}
\description{
\code{fillMs1Spectra} adds empty spectra for cluster that have no MS2 data
}
