% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{getColAnno}
\alias{getColAnno}
\title{Get column annotation data from read .gda file}
\usage{
getColAnno(x)
}
\arguments{
\item{x}{`list` List with data read from .gda file.}
}
\value{
A data frame with the column annotations.
}
\description{
`getColData` returns the actual table with features from the read .gda file
}
\examples{

file <- system.file("extdata/20191108_Pesticides_test_Cluster.gda", package = "genedataRutils")
pest_cluster <- readGda(file)
getColAnno(pest_cluster)

}
\author{
Michael Witting
}
