% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{getMsData}
\alias{getMsData}
\title{Get feature data from read .gda file}
\usage{
getMsData(x)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file.}
}
\value{
A data frame with the feature data
}
\description{
\code{getMsData} returns the actual table with features from the read .gda file
}
\examples{

}
\author{
Michael Witting
}
