% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{getRowAnno}
\alias{getRowAnno}
\title{Get row annotation data from read .gda file}
\usage{
getRowAnno(x)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file.}
}
\value{
A data frame wit row annotations.
}
\description{
\code{getRowData} returns the actual table with features from the read .gda file
}
\examples{

}
\author{
Michael Witting
}
