% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metid.R
\name{matchIonMode}
\alias{matchIonMode}
\title{Compare against a MS2 library}
\usage{
matchIonMode(
  pos,
  neg,
  pos_adducts = c("[M+H]+", "[M+Na]+"),
  neg_adducts = c("[M-H]-"),
  tolerance = 0,
  ppm = 0,
  rtOffset = 0,
  rtimeTolerance = Inf
)
}
\arguments{
\item{pos}{\code{list} List with data read from .gda file containing grouped
MS1 cluster from negative mode}

\item{pos_adducts}{\code{character} vector with adducts in positive mode to check}

\item{neg_adducts}{\code{character} vector with adducts in negative mode to check}

\item{tolerance}{\code{numeric} absolute tolerance for matching of peaks}

\item{ppm}{\code{numeric} relative tolerance for matching of peaks}

\item{rtOffset}{\code{numeric} known offset between measurement and library RT}

\item{rtimeTolerance}{\code{numeric} tolerance for retention time search}
}
\value{
\code{data.frame} with the results
}
\description{
\code{matchIonMode} allows to matches Cluster from different ion modes against each
other. Cluster are matched by retention time and then checked if they share
a common neutral mass.
}
