% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{read_gda_file}
\alias{read_gda_file}
\title{Importing Genedata Expressionist for MS .gda files

Genedata Expressionist for MS can export obtained data in multiple file formats.
One is the .gda format, which can be read by the Analyst module. This file
contains the actual data (e.g. peak intensities, areas etc...) and additional
annotations.
This functions read directly a .gda file and creates three distinct data frames.
One contains the actual MS data (ms_data), the other two the annotations of the columns
and rows (row_anno and col_anno).}
\usage{
read_gda_file(path_to_file, prefix = "")
}
\arguments{
\item{path_to_file}{File path to .gda file}

\item{prefix}{Name prefix for the generated data frames. Should be used if multiple .gda files
shall be read. The prefix is added before each data frame name, e.g. prefix = "cluster" generates
cluster_ms_data as name of data frame variable instead of ms_data. Data frame are automatically assigned
to the current working environment.}
}
\description{
Importing Genedata Expressionist for MS .gda files

Genedata Expressionist for MS can export obtained data in multiple file formats.
One is the .gda format, which can be read by the Analyst module. This file
contains the actual data (e.g. peak intensities, areas etc...) and additional
annotations.
This functions read directly a .gda file and creates three distinct data frames.
One contains the actual MS data (ms_data), the other two the annotations of the columns
and rows (row_anno and col_anno).
}
\examples{
file <- system.file('')
read_gda_file(file, prefix = 'pesticide')

}
