% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sirius.R
\name{read_sirius_result_ms1only}
\alias{read_sirius_result_ms1only}
\title{This file reads the results from SIRIUS back into a data frame}
\usage{
read_sirius_result_ms1only(x)
}
\description{
This file reads the results from SIRIUS back into a data frame
}
