% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{readGda}
\alias{readGda}
\title{Importing Genedata Expressionist for MS .gda files}
\usage{
readGda(file)
}
\arguments{
\item{file}{`character` Path to .gda file that shall be read.}
}
\value{
A list of three data frames, the first contains the actual MS data,
   the second the row annotation and the third the column annotations.
}
\description{
`readGda` Genedata Expressionist for MS can export obtained data in multiple file formats.
One is the .gda format, which can be read by the Analyst module. This file
contains the actual data (e.g. peak intensities, areas etc...) and additional
annotations.
This functions reads directly a .gda file and returns a list with three 
distinct data frames. The first element contains the actual MS data, the 
second contains annotation of the rows and the thrid the annotation of the
columns.
}
\examples{

file <- system.file("extdata/20191108_Pesticides_test_Cluster.gda", package = "genedataRutils")
pest_cluster <- readGda(file)

}
\author{
Michael Witting
}
