% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms1.R
\name{reconstructIsoPattern}
\alias{reconstructIsoPattern}
\title{Importing Genedata Expressionist for MS .gda files}
\usage{
reconstructIsoPattern(peaks, cluster)
}
\arguments{
\item{peaks}{\code{list} List with data read from .gda file containing individual
MS1 peaks}

\item{cluster}{\code{list}List with data read from .gda file containing grouped
MS1 cluster}
}
\value{
A list of three data frames, the first contains the actual MS data,
the second the row annotation and the third the column annotations.
}
\description{
\code{reconstructIsoPattern} Genedata does not allow to export MS1 isotope pattern
directly. They have to be reconstructed from the Peak and Cluster data
}
\examples{


}
\author{
Michael Witting
}
