% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{removeCluster}
\alias{removeCluster}
\title{Remove specific cluster from data}
\usage{
removeCluster(x, cluster)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file.}

\item{cluster}{\code{character} Vector with the cluster to be removed}
}
\description{
\code{removeCluster} Removes specific clusterfrom the data
}
\examples{

}
\author{
Michael Witting
}
