% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{removeSamples}
\alias{removeSamples}
\title{Remove specific samples from data}
\usage{
removeSamples(x, samples)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file.#'}

\item{samples}{\code{character} Vector with the sample names to be removed}
}
\description{
\code{removeSamples} Removes specific samples from the data
}
\examples{

}
\author{
Michael Witting
}
