% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{setColAnno}
\alias{setColAnno}
\title{Sets feature data from read .gda file}
\usage{
setColAnno(x, colAnno)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file.}

\item{colAnno}{\code{data.frame} Replacement for column annotation data in x,
rownames must match}
}
\value{
A list with MS data, row annotations, column annotations
}
\description{
\code{setColAnno} Sets the column annotation data, data is checked if row names match
}
\examples{

}
\author{
Michael Witting
}
