% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{setMsData}
\alias{setMsData}
\title{Sets feature data from read .gda file}
\usage{
setMsData(x, msData)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file.}

\item{msData}{\code{data.frame} Replacement for MS data in x, rownames must match}
}
\value{
A list with MS data, row annotations, column annotations
}
\description{
\code{setMsData} Sets the MS data, data is checked if row names match
}
\examples{

}
\author{
Michael Witting
}
