% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{setRowAnno}
\alias{setRowAnno}
\title{Sets row annotation data from read .gda file}
\usage{
setRowAnno(x, rowAnno)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file.}

\item{rowAnno}{\code{data.frame} Replacement for row annotation data in x,
rownames must match}
}
\value{
A list with MS data, row annotations, column annotations
}
\description{
\code{setRowAnno} Sets the row annotation data, data is checked if row names match
}
\examples{

}
\author{
Michael Witting
}
