% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{writeGda}
\alias{writeGda}
\title{Write .gda file}
\usage{
writeGda(x, file = "export.gda")
}
\arguments{
\item{x}{\code{list} List with data read from .gda file.}

\item{file}{\code{character} Name of file to write to}
}
\description{
\code{writeGda} Writes a list of row and column annotation and ms data to a file
}
\examples{

}
\author{
Michael Witting
}
