% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sirius.R
\name{writeSiriusFile}
\alias{writeSiriusFile}
\title{Importing Genedata Expressionist for MS .gda files}
\usage{
writeSiriusFile(
  ms1_spectra = Spectra(),
  ms2_spectra = Spectra(),
  folder = ".",
  singleFile = FALSE
)
}
\arguments{
\item{ms2_spectra}{\code{Spectra} <code>Spectra</code> containing MS2 data}

\item{folder}{\code{character} path to folder where .ms files shall be stored}

\item{singleFile}{\code{boolean} Indicates if a single file shall be generated,
default is FALSE}

\item{ms1_spectrum}{\code{Spectra} <code>Spectra</code> containing isotope pattern}
}
\description{
\code{writeSiriusFile} Genedata does not allow to export MS1 isotope pattern
directly. They have to be reconstructed from the Peak and Cluster data
}
\examples{


}
\author{
Michael Witting
}
