% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{getColAnno}
\alias{getColAnno}
\title{Get column annotation data from read .gda file}
\usage{
getColAnno(x)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file.}
}
\value{
A data frame with the column annotations.
}
\description{
\code{getColData} returns the actual table with features from the read .gda file
}
\examples{

}
\author{
Michael Witting
}
