% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_gda_files.R
\name{getMsData}
\alias{getMsData}
\title{Get feature data from read .gda file}
\usage{
getMsData(x)
}
\arguments{
\item{x}{`list` List with data read from .gda file.}
}
\value{
A data frame with the feature data
}
\description{
`getMsData` returns the actual table with features from the read .gda file
}
\examples{

file <- system.file("extdata/20191108_Pesticides_test_Cluster.gda", package = "genedataRutils")
pest_cluster <- readGda(file)
getMsData(pest_cluster)

}
\author{
Michael Witting
}
