% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms2.R
\name{ms2AddId}
\alias{ms2AddId}
\title{Importing Genedata Expressionist for MS .gda files}
\usage{
ms2AddId(ms2_spectra, row_anno)
}
\arguments{
\item{peaks}{`list` List with data read from .gda file containing individual 
MS1 peaks}

\item{cluster}{`list`List with data read from .gda file containing grouped 
MS1 cluster}
}
\value{
A list of three data frames, the first contains the actual MS data,
   the second the row annotation and the third the column annotations.
}
\description{
`reconstructIsoPattern` Genedata does not allow to export MS1 isotope pattern
     directly. They have to be reconstructed from the Peak and Cluster data
}
\examples{


}
\author{
Michael Witting
}
