% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ms2.R
\name{ms2AddId}
\alias{ms2AddId}
\title{Importing Genedata Expressionist for MS .gda files}
\usage{
ms2AddId(x, spectra)
}
\arguments{
\item{x}{\code{list} List with data read from .gda file containing grouped
MS1 cluster}

\item{spectra}{\code{Spectra}MSnbase Spectra object containing MS2 spectra to be
matched with the MS1 data}
}
\value{
\code{Spectra} with the additional column CLUSTER_ID.
}
\description{
\code{ms2AddId} Genedata does not allow to export MS1 isotope pattern
directly. They have to be reconstructed from the Peak and Cluster data
}
\examples{


}
\author{
Michael Witting
}
