% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sirius.R
\name{writeSiriusFile}
\alias{writeSiriusFile}
\title{Importing Genedata Expressionist for MS .gda files}
\usage{
writeSiriusFile(ms1_spectra = Spectra(), ms2_spectra = Spectra(), folder = ".")
}
\arguments{
\item{ms2_spectra}{\code{Spectra}}

\item{ms1_spectrum}{\code{Spectra}}

\item{adduct}{\code{character}}

\item{con}{\code{character}}
}
\description{
\code{writeSiriusFile} Genedata does not allow to export MS1 isotope pattern
directly. They have to be reconstructed from the Peak and Cluster data
}
\examples{


}
\author{
Michael Witting
}
